+++ src/style.css
@@ -12,10 +12,6 @@ a {
   color: black;
 }
 
-ul {
-  list-style: none;
-}
-
 /* Styling for Header Starts Here */
 
 header {
@@ -47,12 +43,14 @@ header,
 
 /* Styling for Main section starts Here */
 
-.list {
-  margin: auto;
-  width: 70%;
+ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
   display: flex;
   justify-content: space-evenly;
-  flex-wrap: wrap;
+  justify-content: center;
+  row-gap: 30px;
 }
 
 .list-item {
@@ -75,16 +73,25 @@ header,
   padding: 0.5em 0;
 }
 
-.list-item-likes {
-  text-align: end;
-  padding: 0 0 0.5em 0;
-}
-
 .list-item > button {
   width: 100%;
   padding: 0.5em;
   border: none;
   cursor: pointer;
+}
+
+.list-item-head h4 {
+  margin: 0;
+}
+
+.list-item-head i {
+  margin-left: 10px;
+  color: red;
+}
+
+.list-item-likes {
+  text-align: end;
+  padding: 0 0 0.5em 0;
 }
 
 .list-item > button:hover {
@@ -108,23 +115,35 @@ footer {
 /* Footer styling ends here */
 
 /* popup styling */
-.comments-popup {
+.popup {
   display: none;
   position: fixed;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
+  padding: 20px;
+  margin: 50px auto;
+  text-align: center;
+  background-color: rgba(122, 106, 106, 0.596);
   z-index: 1;
 }
 
-.popup {
+.popup-content {
   background-color: #fff;
   padding: 20px;
   margin: 50px auto;
   max-width: 500px;
   text-align: center;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.popup-img {
+  height: 250px;
+  width: 70%;
+  background-color: #1409468c;
 }
 
 .city-details {
